DocEase/requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
pdf2docx==0.5.8
docx2pdf
pywin32>=306
PyPDF2==3.0.1
Pillow==10.1.0
Werkzeug==2.3.7
WTForms==3.1.1
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3